+++ index.js
@@ -1,3 +1,2 @@
-'use strict';
-
-module.exports = options => require('./build')(options);
+'use strict'
+module.exports = opts => require('./src')(opts)