+++ typings/index.d.ts
@@ -1,4 +1,0 @@
-/// <reference path="globals/handlebars/index.d.ts" />
-/// <reference path="globals/marked/index.d.ts" />
-/// <reference path="globals/mkdirp/index.d.ts" />
-/// <reference path="globals/node/index.d.ts" />