+++ CHANGELOG.md
@@ -1,4 +1,15 @@
-## 18/05/2015
+## 23/05/2015
 
- - Remove IDE links feature
- - Move compiled code to "build" directory
+ - Refactor code base
+ - Removed Typescript
+ - Removed IDE links option
+ - Changed format of options
+ - root option now src
+ - styles is now an array of the stylesheets used by the examples
+ - ext is now an array
+ - colors is now an array
+ - fonts is now an array of objects
+ - scripts is now an array of scripts used by the app
+ - components can now be categorised, using markdown meta
+ - markdown meta now requires a title
+ - order functionality removed from markdown meta